Split game-types.js into node-factories + action-templates; createWAN uses darknet trait
game-types.js held two unrelated concerns under a misleading name (it sat
beside types.js but contained zero typedefs):

- Node-type create* factories (consumed by data/networks/* + mini-network)
- Action templates / ACTION_TEMPLATES + lie-low operator (consumed only by
  traits.js, which already holds the matching timed-action operators)

Renamed game-types.js -> node-factories.js (factories only) and extracted the
action templates into a new action-templates.js. traits.js now imports the
templates from their true home; the dependency flows cleanly
node-factories -> (nothing) and traits -> action-templates.

Also simplified createWAN: it inlined the lie-low operator + darknet/lie-low/
disconnect actions, duplicating exactly what the darknet trait provides. It now
just declares traits: ["darknet"] like every other factory. Its tests resolve
traits before inspecting the (now trait-supplied) operators/actions.

Pure reorganization — no behavior change. make check green (1156 tests).
